+++ lib/flutter_feathersjs.dart
@@ -18,4 +18,5 @@ library flutter_feathersjs;
 
 export './src/config/constants.dart' show FeatherJsErrorType;
 export './src/config/helper.dart' show FeatherJsError;
+export './src/config/storage.dart' show JsonStorage;
 export './src/featherjs.dart';